scripts/document_extraction/requirements.txt
# Document Extraction Requirements - Phase 1 + Phase 3A
# Compatible with Python 3.8+

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF>=1.23.0  # For image extraction from PDFs

# Office documents (DOCX, XLSX, PPTX)
python-docx==0.8.11
openpyxl==3.1.2
python-pptx==0.6.23

# Web documents (HTML, XML)
beautifulsoup4==4.12.2
lxml==4.9.3

# Text encoding detection
chardet==5.2.0

# Text fixing and encoding normalization (Phase 3C)
ftfy>=6.1.1

# Phase 3A: Advanced document analysis
# Language detection
langdetect>=1.0.9

# File type detection (magic bytes)
python-magic>=0.4.27

# Data structure analysis
collections-extended>=2.0.0

# OCR and Image Processing
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0